+++ tests/core/kernel.rs
@@ -18,7 +18,6 @@ fn position() {
 
     let time = spice::str2et("2027-MAR-23 16:00:00");
     let (position, light_time) = spice::spkpos("DIMORPHOS", time, "J2000", "NONE", "HERA");
-    println!("{}", light_time);
 
     for (component, expected_component) in multizip((position.iter(), expected_position.iter())) {
         assert!(relative_eq!(
@@ -42,12 +41,7 @@ fn position() {
 fn already_loaded() -> Result<(), spice::KernelError> {
     let mut kernel = spice::Kernel::new("rsc/krn/hera_study_PO_EMA_2024.tm")?;
 
-    assert_eq!(
-        kernel.load(),
-        Err(spice::KernelError {
-            kind: spice::KernelErrorKind::AlreadyLoaded
-        })
-    );
+    assert!(kernel.load().is_err());
 
     kernel.unload()?;
     Ok(())
@@ -55,29 +49,12 @@ fn already_loaded() -> Result<(), spice::KernelError> {
 
 #[test]
 #[serial]
-fn display() -> Result<(), spice::KernelError> {
+fn already_unloaded() -> Result<(), spice::KernelError> {
     let mut kernel = spice::Kernel::new("rsc/krn/hera_study_PO_EMA_2024.tm")?;
 
-    match kernel.load() {
-        Ok(_) => (),
-        Err(e) => assert_eq!(e.to_string(), "the kernel is already loaded"),
-    };
-
     kernel.unload()?;
-    Ok(())
-}
 
-#[test]
-#[serial]
-fn debug() -> Result<(), spice::KernelError> {
-    let mut kernel = spice::Kernel::new("rsc/krn/hera_study_PO_EMA_2024.tm")?;
-
-    match kernel.load() {
-        Ok(_) => (),
-        Err(e) => assert_eq!(format!("{:?}", e), r#""the kernel is already loaded""#),
-    };
-
-    kernel.unload()?;
+    assert!(kernel.unload().is_err());
     Ok(())
 }