+++ CHANGELOG.md
@@ -12,16 +12,41 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.8] - 2021-05-12
+
+### Added
+
++ implement occult_c
+
+### Changed
+
++ improve readme
++ simplify error handling with crate `thiserror`
+
+### Removed
+
++ unnecessary examples
++ unnecessary tests
+
 ## [0.4.7] - 2021-04-23
+
+### Changed
 
 + using nalgebra 0.26
 
 ## [0.4.6] - 2021-04-16
 
+### Added
+
 + more examples
+
+### Changed
+
 + use rustool 0.3.12
 
 ## [0.4.5] - 2021-04-09
+
+### Changed
 
 + improved example